# docs/architecture/openapi.phase1.yaml
openapi: 3.0.3
info:
  title: PawMate Phase 1 API
  version: 1.0.0
  description: Phase 1 contract for auth, pets, vets, reviews, health records, reminders, and notifications.
servers:
  - url: http://localhost:3000
  - url: https://api.pawmate.example
...
security:
  - bearerAuth: []
paths:
...
  /pets:
    get:
      tags: [Pets]
      summary: List current user's pets
      responses:
        '200':
          description: Pet list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetList'
    post:
      tags: [Pets]
      summary: Create a pet profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetCreateRequest'
      responses:
        '201':
          description: Pet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetItem'
  /pets/{petId}:
    parameters:
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [Pets]
      summary: Get a pet profile
      responses:
        '200':
          description: Pet profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetItem'
    put:
      tags: [Pets]
      summary: Update a pet profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetUpdateRequest'
      responses:
        '200':
          description: Updated pet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetItem'
    delete:
      tags: [Pets]
      summary: Soft delete a pet profile
      responses:
        '204':
          description: Pet removed
  /pets/{petId}/photo:
    parameters:
      - $ref: '#/components/parameters/PetId'
    post:
      tags: [Pets]
      summary: Upload or replace a pet avatar photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetPhotoUploadRequest'
      responses:
        '200':
          description: Pet photo stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetPhotoUploadResponse'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
  /pets/{petId}/reminders:
    parameters:
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [Reminders]
      summary: List reminders for a pet
      responses:
        '200':
          description: Reminder list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderList'
    post:
      tags: [Reminders]
      summary: Create a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderUpsert'
      responses:
        '201':
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderItem'
  /pets/{petId}/reminders/{reminderId}:
    parameters:
      - $ref: '#/components/parameters/PetId'
      - $ref: '#/components/parameters/ReminderId'
    get:
      tags: [Reminders]
      summary: Get a reminder
      responses:
        '200':
          description: Reminder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderItem'
    put:
      tags: [Reminders]
      summary: Update a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderUpsert'
      responses:
        '200':
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderItem'
    delete:
      tags: [Reminders]
      summary: Delete a reminder
      responses:
        '204':
          description: Reminder removed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PetId:
      in: path
      name: petId
      required: true
      schema: { type: string, format: uuid }
...
    ReminderId:
      in: path
      name: reminderId
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [success, error, requestId]
      additionalProperties: false
      properties:
        success:
          type: boolean
          enum: [false]
        requestId: { type: string }
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message: { type: string }
            field: { type: string, nullable: true }
    ErrorCode:
      oneOf:
        - $ref: '#/components/schemas/AuthErrorCode'
        - $ref: '#/components/schemas/PetErrorCode'
        - $ref: '#/components/schemas/VetErrorCode'
    AuthErrorCode:
      type: string
      enum:
        - AUTH_001
        - AUTH_002
        - AUTH_003
        - AUTH_004
        - AUTH_005
        - AUTH_006
        - AUTH_007
        - AUTH_008
        - AUTH_009
        - AUTH_010
        - AUTH_011
        - AUTH_012
        - AUTH_013
        - AUTH_014
        - AUTH_015
        - AUTH_016
        - AUTH_017
        - AUTH_018
        - AUTH_019
        - AUTH_020
    PetErrorCode:
      type: string
      enum:
        - PET_001
        - PET_002
        - PET_003
        - PET_004
        - PET_005
        - PET_006
        - PET_007
        - PET_008
        - PET_009
        - PET_010
    VetErrorCode:
      type: string
      enum:
        - VET_001
        - VET_002
        - VET_003
        - VET_004
        - VET_005
        - VET_006
        - VET_007
        - VET_008
        - VET_009
        - VET_010
    Meta:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    Pet:
      type: object
      required: [id, userId, name, species, gender]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        name: { type: string }
        species: { $ref: '#/components/schemas/PetSpecies' }
        breed: { type: string, nullable: true }
        gender: { $ref: '#/components/schemas/PetGender' }
        dob: { type: string, format: date, nullable: true }
        weight: { type: number, format: float, nullable: true }
        color: { type: string, nullable: true }
        microchip: { type: string, nullable: true }
        isNeutered: { type: boolean }
        avatarUrl: { type: string, format: uri, nullable: true }
        healthStatus: { $ref: '#/components/schemas/PetHealthStatus' }
    PetSpecies:
      type: string
      enum: [dog, cat, bird, rabbit, other]
    PetGender:
      type: string
      enum: [male, female, unknown]
    PetHealthStatus:
      type: string
      enum: [healthy, monitoring, chronic, recovery, unknown]
    PetCreateRequest:
      type: object
      required: [name, species, gender]
      properties:
        name: { type: string }
        species: { $ref: '#/components/schemas/PetSpecies' }
        breed: { type: string, nullable: true }
        gender: { $ref: '#/components/schemas/PetGender' }
        dob: { type: string, format: date, nullable: true }
        weight: { type: number, format: float, nullable: true }
        color: { type: string, nullable: true }
        microchip: { type: string, nullable: true }
        isNeutered: { type: boolean, default: false }
        avatarUrl: { type: string, format: uri, nullable: true }
        healthStatus: { $ref: '#/components/schemas/PetHealthStatus' }
    PetUpdateRequest:
      type: object
      properties:
        name: { type: string }
        species: { $ref: '#/components/schemas/PetSpecies' }
        breed: { type: string, nullable: true }
        gender: { $ref: '#/components/schemas/PetGender' }
        dob: { type: string, format: date, nullable: true }
        weight: { type: number, format: float, nullable: true }
        color: { type: string, nullable: true }
        microchip: { type: string, nullable: true }
        isNeutered: { type: boolean }
        avatarUrl: { type: string, format: uri, nullable: true }
        healthStatus: { $ref: '#/components/schemas/PetHealthStatus' }
    PetPhotoUploadRequest:
      type: object
      required: [fileName, contentType, base64Data]
      properties:
        fileName:
          type: string
          example: milo.png
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        base64Data:
          type: string
          description: Base64-encoded image body without transport headers.
    PetPhotoUploadResponse:
      type: object
      required: [avatarUrl, pet]
      properties:
        avatarUrl:
          type: string
          format: uri
        pet:
          $ref: '#/components/schemas/PetItem'
    PetItem:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Pet' }
    PetList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Pet' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    Reminder:
      type: object
      required: [id, petId, title, reminderAt, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        petId: { type: string, format: uuid }
        title: { type: string }
        reminderAt: { type: string, format: date-time }
        repeatRule: { type: string, nullable: true }
        status: { type: string, enum: [active, paused, completed, cancelled] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ReminderUpsert:
      type: object
      required: [title, reminderAt]
      properties:
        title: { type: string }
        reminderAt: { type: string, format: date-time }
        repeatRule: { type: string, nullable: true }
        status: { type: string, enum: [active, paused, completed, cancelled] }
    ReminderItem:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Reminder' }
    ReminderList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Reminder' }
        meta: { $ref: '#/components/schemas/Meta' }
...